+++ src/motion/features/types.ts
@@ -1,7 +1,5 @@
 import * as React from "react"
 import { MotionProps } from "../types"
-import { VisualElementAnimationControls } from "../../animation/VisualElementAnimationControls"
-import { MotionContextProps } from "../context/MotionContext"
 import { MotionConfigContext } from "../context/MotionConfigContext"
 import { VisualElement } from "../../render/VisualElement"
 import { HTMLVisualElement } from "../../render/dom/HTMLVisualElement"
@@ -11,9 +9,6 @@ import { HTMLVisualElement } from "../../render/dom/HTMLVisualElement"
  */
 export interface FeatureProps extends MotionProps {
     visualElement: HTMLVisualElement
-    controls: VisualElementAnimationControls
-    localContext: MotionContextProps
-    parentContext: MotionContextProps
 }
 
 /**
@@ -21,10 +16,7 @@ export interface FeatureProps extends MotionProps {
  */
 export interface MotionFeature {
     key: string
-    shouldRender: (
-        props: MotionProps,
-        parentContext: MotionContextProps
-    ) => boolean
+    shouldRender: (props: MotionProps) => boolean
     getComponent: (
         props: MotionProps
     ) => React.ComponentType<FeatureProps> | undefined
@@ -33,9 +25,6 @@ export interface MotionFeature {
 export type LoadMotionFeatures<P = {}> = (
     visualElenent: VisualElement,
     props: P & MotionProps,
-    context: MotionContextProps,
-    parentContext: MotionContextProps,
-    controls: VisualElementAnimationControls<P>,
     inherit: boolean,
     plugins: MotionConfigContext
 ) => React.ReactElement<FeatureProps>[]