+++ src/motion/index.tsx
@@ -1,22 +1,20 @@
 import * as React from "react"
-import { useContext, forwardRef, Ref } from "react"
-import { MotionContext, useMotionContext } from "./context/MotionContext"
+import { forwardRef, Ref, useContext, useMemo } from "react"
 import { MotionProps } from "./types"
-import { checkShouldInheritVariant } from "./utils/should-inherit-variant"
 import { useMotionValues } from "./utils/use-motion-values"
-import { UseVisualElement } from "../render/types"
+import { UseVisualElement } from "../render/VisualElement/types"
 import { RenderComponent, MotionFeature } from "./features/types"
-import { AnimationControlsConfig } from "../animation/VisualElementAnimationControls"
-import { useVisualElementAnimation } from "../animation/use-visual-element-animation"
 import { useFeatures } from "./features/use-features"
 import { useSnapshotOnUnmount } from "./features/layout/use-snapshot-on-unmount"
+import { useVariants } from "./utils/use-variants"
+import { MotionConfigContext } from "./context/MotionConfigContext"
+import { MotionContext } from "./context/MotionContext"
 export { MotionProps }
 
 export interface MotionComponentConfig<E> {
     defaultFeatures: MotionFeature[]
     useVisualElement: UseVisualElement<E>
     render: RenderComponent
-    animationControlsConfig: AnimationControlsConfig
 }
 
 /**
@@ -32,24 +30,16 @@ export interface MotionComponentConfig<E> {
  */
 export function createMotionComponent<P extends {}, E>(
     Component: string | React.ComponentType<P>,
-    {
-        defaultFeatures,
-        useVisualElement,
-        render,
-        animationControlsConfig,
-    }: MotionComponentConfig<E>
+    { defaultFeatures, useVisualElement, render }: MotionComponentConfig<E>
 ) {
     function MotionComponent(props: P & MotionProps, externalRef?: Ref<E>) {
-        const parentContext = useContext(MotionContext)
-        const shouldInheritVariant = checkShouldInheritVariant(props)
-
         /**
-         * If a component isStatic, we only visually update it as a
+         * If a component is static, we only visually update it as a
          * result of a React re-render, rather than any interactions or animations.
-         * If this component or any ancestor isStatic, we disable hardware acceleration
+         * If this component or any ancestor is static, we disable hardware acceleration
          * and don't load any additional functionality.
          */
-        const isStatic = parentContext.static || props.static || false
+        const { isStatic } = useContext(MotionConfigContext)
 
         /**
          * Create a VisualElement for this component. A VisualElement provides a common
@@ -60,38 +50,19 @@ export function createMotionComponent<P extends {}, E>(
         const visualElement = useVisualElement(
             Component,
             props,
-            parentContext.visualElement as any,
             isStatic,
             externalRef
         )
 
         /**
-         * Scrape MotionValues from props and add/remove them to/from
-         * the VisualElement as necessary.
+         * Scrape MotionValues from props and add/remove them to/from the VisualElement.
          */
         useMotionValues(visualElement, props)
 
         /**
-         * Create animation controls for the VisualElement. It might be
-         * interesting to try and combine this with VisualElement itself in a further refactor.
+         * Add the visualElement as a node in the variant tree.
          */
-        const controls = useVisualElementAnimation(
-            visualElement,
-            props,
-            animationControlsConfig,
-            shouldInheritVariant
-        )
-
-        /**
-         * Build the MotionContext to pass on to the next `motion` component.
-         */
-        const context = useMotionContext(
-            parentContext,
-            controls,
-            visualElement,
-            isStatic,
-            props
-        )
+        const variantContext = useVariants(visualElement, props, isStatic)
 
         /**
          * Load features as renderless components unless the component isStatic
@@ -100,12 +71,16 @@ export function createMotionComponent<P extends {}, E>(
             defaultFeatures,
             isStatic,
             visualElement,
-            controls,
-            props,
-            context,
-            parentContext,
-            shouldInheritVariant
+            props
         )
+
+        /**
+         * Only create a new context value when the sub-contexts change.
+         */
+        const context = useMemo(() => ({ visualElement, variantContext }), [
+            visualElement,
+            variantContext,
+        ])
 
         const component = render(Component, props, visualElement)