+++ src/render/VisualElement/types.ts
@@ -1,5 +1,5 @@
-import { VisualElement } from "./VisualElement"
-import { MotionProps } from "../motion/types"
+import { VisualElement } from "."
+import { MotionProps } from "../../motion/types"
 import { Ref } from "react"
 
 /**
@@ -10,7 +10,6 @@ import { Ref } from "react"
 export type UseVisualElement<E, P = MotionProps> = (
     Component: string | React.ComponentType<P>,
     props: MotionProps & P,
-    parent?: VisualElement,
     isStatic?: boolean,
     ref?: Ref<E>
 ) => VisualElement