+++ src/render/dom/use-dom-visual-element.ts
@@ -2,10 +2,11 @@ import { HTMLVisualElement } from "./HTMLVisualElement"
 import { useConstant } from "../../utils/use-constant"
 import { MotionProps } from "../../motion/types"
 import { SVGVisualElement } from "./SVGVisualElement"
-import { UseVisualElement } from "../types"
+import { UseVisualElement } from "../VisualElement/types"
 import { isSVGComponent } from "./utils/is-svg-component"
-import { useIsPresent } from "../../components/AnimatePresence/use-presence"
-import { useEffect } from "react"
+import { useContext, useEffect } from "react"
+import { PresenceContext } from "../../components/AnimatePresence/PresenceContext"
+import { useVisualElementContext } from "../../motion/context/MotionContext"
 
 /**
  * DOM-flavoured variation of the useVisualElement hook. Used to create either a HTMLVisualElement
@@ -15,10 +16,11 @@ import { useEffect } from "react"
 export const useDomVisualElement: UseVisualElement<MotionProps, any> = (
     Component,
     props,
-    parent,
     isStatic,
     ref
 ) => {
+    const parent = useVisualElementContext()
+
     const visualElement = useConstant(() => {
         const DOMVisualElement = isSVGComponent(Component)
             ? SVGVisualElement
@@ -35,10 +37,25 @@ export const useDomVisualElement: UseVisualElement<MotionProps, any> = (
 
     visualElement.layoutId = props.layoutId
 
-    const isPresent = useIsPresent()
+    const presenceContext = useContext(PresenceContext)
+
+    /**
+     * Update VisualElement with presence data.
+     */
+    const isPresent =
+        presenceContext === null ? true : presenceContext.isPresent
     visualElement.isPresent =
         props.isPresent !== undefined ? props.isPresent : isPresent
 
+    /**
+     *
+     */
+    const presenceId = presenceContext?.id
+    visualElement.isPresenceRoot = !parent || parent.presenceId !== presenceId
+
+    /**
+     * TODO: Investigate if we need this
+     */
     useEffect(() => {
         if (props.onViewportBoxUpdate) {
             return visualElement.onViewportBoxUpdate(props.onViewportBoxUpdate)